+++ examples/admin-sdk-app/src/frontend/locations/ex-chart-card-before.vue
@@ -22,8 +22,8 @@
 </template>
 
 <script setup lang="ts">
-import { window, context } from "@shopware-ag/meteor-admin-sdk";
-import { SwButton } from '@shopware-ag/meteor-component-library';
+import { window, context } from "@allincart-ag/meteor-admin-sdk";
+import { SwButton } from '@allincart-ag/meteor-component-library';
 
 async function goToExampleModule() {
   const moduleInformation = await context.getModuleInformation();