+++ packages/component-library/src/components/form/mt-url-field/mt-url-field.vue
@@ -248,10 +248,10 @@ const { copy, copied } = useClipboard();
 
 const { t } = useI18n({
   messages: {
-    de: {
-      copyTooltip: "URL in Zwischenablage kopieren",
-      copyButtonDescription: "URL in Zwischenablage kopieren",
-      copyButtonDescriptionValueCopied: "URL in Zwischenablage kopiert",
+    zh: {
+      "copyTooltip": "复制 URL 到剪贴板",
+      "copyButtonDescription": "复制 URL 到剪贴板",
+      "copyButtonDescriptionValueCopied": "URL 已复制到剪贴板"
     },
     en: {
       copyTooltip: "Copy URL to clipboard",