+++ packages/component-library/src/components/icons-media/mt-icon/mt-icon.vue
@@ -86,7 +86,7 @@ watch(
   iconInformation,
   () => {
     import(
-      `./../../../../node_modules/@shopware-ag/meteor-icon-kit/icons/${iconInformation.value.mode}/${iconInformation.value.name}.svg`
+      `./../../../../node_modules/@allincart-ag/meteor-icon-kit/icons/${iconInformation.value.mode}/${iconInformation.value.name}.svg`
     ).then((data) => {
       if (data.default) {
         iconSvgData.value = data.default;
@@ -101,7 +101,7 @@ watch(
 </script>
 
 <style lang="scss">
-@import "node_modules/@shopware-ag/meteor-icon-kit/icons/meteor-icon-kit.scss";
+@import "node_modules/@allincart-ag/meteor-icon-kit/icons/meteor-icon-kit.scss";
 
 .mt-icon {
   display: inline-block;