+++ packages/component-library/src/components/table-and-list/mt-pagination/mt-pagination.vue
@@ -94,12 +94,12 @@ const { t } = useI18n({
       nextPage: "Next page",
       lastPage: "Last page",
     },
-    de: {
-      infoText: "{start}-{end} von {totalItems}",
-      firstPage: "Erste Seite",
-      previousPage: "Voherige Seite",
-      nextPage: "Nächste Seite",
-      lastPage: "Letzte Seite",
+    zh: {
+      infoText: "{start}-{end} 共 {totalItems}",
+      firstPage: "第一页",
+      previousPage: "上一页",
+      nextPage: "下一页",
+      lastPage: "最后一页",
     },
   },
 });
@@ -222,6 +222,7 @@ const lastVisibleItemNumber = computed(() => {
   }
 
   /* Firefox */
+
   &[type="number"] {
     -moz-appearance: textfield;
   }